commands: move split code to split.rs
